pdfBot/requirements.txt
langchain==0.1.4
langchain-community==0.0.16
langchain-core==0.1.23
langchain-openai==0.0.5
langchain-pinecone==0.0.3
python-dotenv==1.0.1
streamlit==1.30.0
beautifulsoup4==4.12.2
PyPDF2==3.0.1
altair==5.3.0
tiktoken==0.5.2
InstructorEmbedding==1.0.1
sentence_transformers==3.0.1
